Fix: display an experimental vault in the vault page.
