+++ src/utils/strategies.ts
@@ -19,7 +19,7 @@ import { mapContractCalls } from './commonUtils';
 
 import StratABI from './ABI/Strategy.json';
 import TokenABI from './ABI/Token.json';
-import { getVaults } from './vaults';
+import { getEndorsedVaults } from './vaults';
 
 interface VaultVersionInfo {
     apiVersion: string;
@@ -255,7 +255,7 @@ const innerGetStrategies = async (addresses: string[]): Promise<Strategy[]> => {
 export const getStrategies = memoize(innerGetStrategies);
 
 export const _getAllStrategies = async (): Promise<Strategy[]> => {
-    const allVaults = await getVaults();
+    const allVaults = await getEndorsedVaults();
     return allVaults.flatMap((vault) => vault.strategies);
 };
