+++ src/utils/vaults.ts
@@ -7,33 +7,10 @@ import { BigNumber, utils } from 'ethers';
 import { BigNumber as BigNumberJS } from 'bignumber.js';
 import { get, memoize } from 'lodash';
 import compareVersions from 'compare-versions';
-import { getEthersDefaultProvider } from './ethers';
-import { Vault, VaultApi, VaultVersion, Strategy } from '../types';
-import { BuildGet } from './apisRequest';
-import { vaultChecks } from './checks';
-import {
-    mapContractCalls,
-    createStrategiesHelperCallAssetStrategiesAddresses,
-    mapToStrategyAddressQueueIndex,
-} from './commonUtils';
-import { toHumanDateText } from './dateUtils';
-import { getABI_032 } from './abi';
-import { mapStrategiesCalls, buildStrategyCalls } from './strategies';
-import { getTotalDebtUsage } from './strategyParams';
-
-const VAULT_VIEW_METHODS = [
-    'management',
-    'managementFee',
-    'performanceFee',
-    'governance',
-    'guardian',
-    'depositLimit',
-    'totalAssets',
-    'debtRatio',
-    'totalDebt',
-    'lastReport',
-    'rewards',
-];
+import { Vault, VaultApi, VaultVersion } from '../types';
+import { BuildGet, VAULTS_ALL, VAULTS_ALL_EXPERIMENTAL } from './apisRequest';
+import { DEFAULT_QUERY_PARAM, QueryParam } from './types/QueryParam';
+import { fillVaultData, mapVaultDataToVault } from './vaultDataMapping';
 
 interface VaultData {
     apiVersion: string;
@@ -85,150 +62,100 @@ const filterAndMapVaultsData = (
                     !FILTERED_VAULTS.has(vault.address.toLowerCase())) ||
                 filterList.has(vault.address.toLowerCase())
         )
-        .map((vault: any) => {
-            return {
-                ...vault,
-                apiVersion: vault.version,
-                name: vault.display_name,
-                emergencyShutdown: vault.emergency_shutdown,
-                tvl: {
-                    totalAssets: BigNumber.from(
-                        new BigNumberJS(
-                            vault.tvl.total_assets.toString()
-                        ).toFixed(0)
-                    ),
-                },
-            } as VaultApi;
-        });
+        .map(fillVaultData);
     // DEV NOTE: this is a helper method from debug.ts for debugging the data, should do nothing in prod
     // vaultData = debugFilter(vaultData);
 
     return vaultData;
 };
 
-const vaultsAreMissing = (
-    vaultMap: Map<string, VaultApi>,
-    additional: Set<string>
-): boolean => {
-    let missing = false;
-    additional.forEach((vaultAddr) => {
-        if (vaultMap.has(vaultAddr.toLowerCase()) === false) {
-            missing = true;
-        }
-    });
-
-    return missing;
-};
-
 export const getTotalVaults = async (): Promise<number> => {
-    const response = await BuildGet('/vaults/all');
+    const response = await BuildGet(VAULTS_ALL);
 
     const payload: VaultApi[] = filterAndMapVaultsData(response.data);
 
     return payload.length;
 };
 
-const _internalGetVaults = async (
+const _getEndorsedVaults = async (
     allowList: string[] = [],
-    offset = 0,
-    limit = 200
+    queryParams: QueryParam = DEFAULT_QUERY_PARAM
 ): Promise<Vault[]> => {
-    const provider = getEthersDefaultProvider();
-
-    const multicall = new Multicall({ ethersProvider: provider });
     // accepts non endorsed experimental vaults to access
     const filterList = new Set(allowList.map((addr) => addr.toLowerCase()));
 
-    const response = await BuildGet('/vaults/all');
+    const response = await BuildGet(VAULTS_ALL);
     const sortedVaultList = sortVaultsByVersion(response.data);
     let payload: VaultApi[] = filterAndMapVaultsData(
         sortedVaultList,
         filterList
     );
 
-    if (offset >= payload.length) {
+    if (queryParams.pagination.offset >= payload.length) {
         return [];
     }
     payload = payload.slice(
-        Math.max(0, offset),
-        Math.min(payload.length, offset + limit)
+        Math.max(0, queryParams.pagination.offset),
+        Math.min(
+            payload.length,
+            queryParams.pagination.offset + queryParams.pagination.limit
+        )
     );
-    const vaultMap = new Map<string, VaultApi>();
-    const strategyMap = new Map<string, string>();
+    const vaults: Vault[] = await mapVaultDataToVault(payload);
 
-    payload.forEach((vault) => {
-        vaultMap.set(vault.address, vault);
-        vault.strategies.forEach((strat) =>
-            strategyMap.set(strat.address, vault.address)
-        );
-    });
-
-    // TODO: uncomment and improve this
-    // // check if we have missing vaults from requested
-    // if (vaultsAreMissing(vaultMap, additional)) {
-    //     // need to fetch experimental data
-    //     console.log('...fetching experimental vaults data');
-    //     const response = await BuildGetExperimental('/vaults/all');
-    //     const experimentalPayload: VaultApi[] = filterAndMapVaultsData(
-    //         response.data,
-    //         additional
-    //     );
-    //     experimentalPayload.forEach((vault) => {
-    //         vaultMap.set(vault.address, vault);
-    //         vault.strategies.forEach((strat) =>
-    //             strategyMap.set(strat.address, vault.address)
-    //         );
-    //     });
-    // }
-
-    const vaultCalls: ContractCallContext[] = payload.map(({ address }) => {
-        const calls = VAULT_VIEW_METHODS.map((method) => ({
-            reference: method,
-            methodName: method,
-            methodParameters: [],
-        }));
-        return {
-            reference: address,
-            contractAddress: address,
-            abi: getABI_032(), // only this abi version has the vault view methods
-            calls,
-        };
-    });
-    const stratCalls: ContractCallContext[] = payload.flatMap(
-        ({ strategies }) => {
-            const stratAddresses = strategies.map(({ address }) => address);
-            return buildStrategyCalls(stratAddresses, vaultMap, strategyMap);
-        }
-    );
-    const strategiesHelperCallResults: ContractCallResults = await multicall.call(
-        createStrategiesHelperCallAssetStrategiesAddresses(payload)
-    );
-    const results: ContractCallResults = await multicall.call(
-        vaultCalls.concat(stratCalls)
-    );
-
-    return mapVaultData(
-        results,
-        strategiesHelperCallResults,
-        vaultMap,
-        strategyMap
-    );
+    return vaults;
 };
-export const getVaultsWithPagination = memoize((offset = 0, limit = 100) =>
-    _internalGetVaults([], offset, limit)
-);
-export const getVaults = memoize(_internalGetVaults);
 
-const _getVault = async (address: string): Promise<Vault> => {
+const _getExperimentalVaults = async (
+    allowList: string[] = [],
+    queryParams: QueryParam = DEFAULT_QUERY_PARAM
+): Promise<Vault[]> => {
+    // accepts non endorsed experimental vaults to access
+    const filterList = new Set(allowList.map((addr) => addr.toLowerCase()));
+
+    const response = await BuildGet(VAULTS_ALL_EXPERIMENTAL);
+    const sortedVaultList = sortVaultsByVersion(response.data);
+    let payload: VaultApi[] = filterAndMapVaultsData(
+        sortedVaultList,
+        filterList
+    );
+
+    if (queryParams.pagination.offset >= payload.length) {
+        return [];
+    }
+    payload = payload.slice(
+        Math.max(0, queryParams.pagination.offset),
+        Math.min(
+            payload.length,
+            queryParams.pagination.offset + queryParams.pagination.limit
+        )
+    );
+    const vaults: Vault[] = await mapVaultDataToVault(payload);
+
+    return vaults;
+};
+
+export const getVaultsWithPagination = memoize(
+    (queryParams: QueryParam = DEFAULT_QUERY_PARAM) =>
+        _getEndorsedVaults([], queryParams)
+);
+export const getEndorsedVaults = memoize(_getEndorsedVaults);
+
+const _getEndorsedOrExperimentalVault = async (
+    address: string
+): Promise<Vault> => {
     if (!address || !utils.isAddress(address)) {
         throw new Error('Expected a valid vault address');
     }
 
-    const vaults = await getVaults([address]);
+    const vaults = await getEndorsedVaults([address]);
 
-    const [foundVault]: Vault[] = vaults.filter(
+    let [foundVault]: Vault[] = vaults.filter(
         (vault) => vault.address.toLowerCase() === address.toLowerCase()
     );
+    if (!foundVault) {
+        [foundVault] = await _getExperimentalVaults([address]);
+    }
 
     if (!foundVault) {
         throw new Error('Requested address not recognized as a yearn vault');
@@ -237,74 +164,6 @@ const _getVault = async (address: string): Promise<Vault> => {
     return foundVault;
 };
 
-export const getVault = memoize(_getVault);
-
-const mapVaultData = (
-    contractCallsResults: ContractCallResults,
-    strategiesHelperCallsResults: ContractCallResults,
-    vaultMap: Map<string, VaultApi>,
-    strategyMap: Map<string, string>
-): Vault[] => {
-    const vaults: Vault[] = [];
-
-    vaultMap.forEach((vault, key) => {
-        const {
-            address,
-            apiVersion,
-            symbol,
-            name,
-            token,
-            icon,
-            emergencyShutdown,
-            strategies,
-        } = vault;
-
-        const strategiesQueueIndexes = mapToStrategyAddressQueueIndex(
-            address,
-            strategiesHelperCallsResults
-        );
-
-        const mappedVault: any = {
-            address,
-            apiVersion,
-            symbol,
-            name,
-            token,
-            icon,
-            emergencyShutdown,
-        };
-
-        const stratAddresses = strategies.map(({ address }) => address);
-        const mappedStrategies: Strategy[] = mapStrategiesCalls(
-            stratAddresses,
-            contractCallsResults,
-            strategiesQueueIndexes,
-            strategyMap
-        );
-
-        mappedVault.debtUsage = getTotalDebtUsage(mappedStrategies);
-
-        const vaultData = contractCallsResults.results[address];
-
-        const mappedVaultContractCalls = mapContractCalls(vaultData);
-        const mappedVaultContractCallsConverted = {
-            ...mappedVaultContractCalls,
-            managementFee: parseInt(mappedVaultContractCalls.managementFee),
-            performanceFee: parseInt(mappedVaultContractCalls.performanceFee),
-        };
-
-        mappedVault.lastReportText = toHumanDateText(
-            mappedVaultContractCalls.lastReport
-        );
-
-        vaults.push(
-            vaultChecks({
-                ...mappedVault,
-                ...mappedVaultContractCallsConverted,
-                strategies: mappedStrategies,
-            })
-        );
-    });
-
-    return vaults;
-};
+export const getEndorsedOrExperimentalVault = memoize(
+    _getEndorsedOrExperimentalVault
+);
